+++ tests/spcs/test_services.py
@@ -314,11 +314,12 @@ def test_deploy_service(
         IN COMPUTE POOL test_compute_pool
         FROM @test_stage
         SPECIFICATION_FILE = 'spec.yml'
+        AUTO_RESUME = False
         MIN_INSTANCES = 1
         MAX_INSTANCES = 2
         QUERY_WAREHOUSE = xsmall
-        COMMENT = 'This is a test service'
         EXTERNAL_ACCESS_INTEGRATIONS = (test_external_access_integration)
+        COMMENT = 'This is a test service'
         WITH TAG (test_tag='test_value')"""
         )
         assert result.exit_code == 0, result.output
@@ -331,59 +332,6 @@ def test_deploy_service(
                     f"put file://{Path(tmp_dir).resolve() / 'output' / 'bundle' / 'service' / 'spec.yml'} @test_stage auto_compress=false parallel=4 overwrite=True",
                     cursor_class=SnowflakeCursor,
                 ),
-            ]
-        )
-
-
-@patch("snowflake.cli._plugins.object.manager.ObjectManager.execute_query")
-@patch("snowflake.cli._plugins.stage.manager.StageManager.execute_query")
-@patch(EXECUTE_QUERY)
-def test_deploy_service_replace(
-    mock_execute_query,
-    mock_stage_manager_execute_query,
-    mock_object_manager_execute_query,
-    runner,
-    project_directory,
-    mock_cursor,
-    os_agnostic_snapshot,
-):
-    mock_execute_query.return_value = mock_cursor(
-        rows=[["Service TEST_SERVICE successfully created."]],
-        columns=["status"],
-    )
-
-    with project_directory("spcs_service") as tmp_dir:
-        result = runner.invoke(["spcs", "service", "deploy", "--replace"])
-
-        expected_query = dedent(
-            """\
-        CREATE SERVICE test_service
-        IN COMPUTE POOL test_compute_pool
-        FROM @test_stage
-        SPECIFICATION_FILE = 'spec.yml'
-        MIN_INSTANCES = 1
-        MAX_INSTANCES = 2
-        QUERY_WAREHOUSE = xsmall
-        COMMENT = 'This is a test service'
-        EXTERNAL_ACCESS_INTEGRATIONS = (test_external_access_integration)
-        WITH TAG (test_tag='test_value')"""
-        )
-        assert result.exit_code == 0, result.output
-        assert result.output == os_agnostic_snapshot
-        mock_stage_manager_execute_query.assert_has_calls(
-            [
-                call("create stage if not exists IDENTIFIER('test_stage')"),
-                call(
-                    f"put file://{Path(tmp_dir).resolve() / 'output' / 'bundle' / 'service' / 'spec.yml'} @test_stage auto_compress=false parallel=4 overwrite=True",
-                    cursor_class=SnowflakeCursor,
-                ),
-            ]
-        )
-        mock_execute_query.assert_called_once_with(expected_query)
-        mock_object_manager_execute_query.assert_has_calls(
-            [
-                call("describe service IDENTIFIER('test_service')"),
-                call("drop service IDENTIFIER('test_service')"),
             ]
         )
 
@@ -415,11 +363,12 @@ def test_deploy_service_already_exists(
         IN COMPUTE POOL test_compute_pool
         FROM @test_stage
         SPECIFICATION_FILE = 'spec.yml'
+        AUTO_RESUME = False
         MIN_INSTANCES = 1
         MAX_INSTANCES = 2
         QUERY_WAREHOUSE = xsmall
-        COMMENT = 'This is a test service'
         EXTERNAL_ACCESS_INTEGRATIONS = (test_external_access_integration)
+        COMMENT = 'This is a test service'
         WITH TAG (test_tag='test_value')"""
         )
         assert result.exit_code == 1, result.output
@@ -476,11 +425,12 @@ def test_deploy_multiple_services(
         IN COMPUTE POOL test_compute_pool
         FROM @test_stage
         SPECIFICATION_FILE = 'spec.yml'
+        AUTO_RESUME = True
         MIN_INSTANCES = 1
         MAX_INSTANCES = 2
         QUERY_WAREHOUSE = xsmall
-        COMMENT = 'This is a test service'
         EXTERNAL_ACCESS_INTEGRATIONS = (test_external_access_integration)
+        COMMENT = 'This is a test service'
         WITH TAG (test_tag='test_value')"""
         )
         assert result.exit_code == 0, result.output